requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pygame==2.6.1
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
